backend/requirements.txt
# FastAPI & Web Server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0

# Embeddings & Vector Search
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# Local LLM (Llama.cpp)
llama-cpp-python==0.2.50

# Extra Utilities
numpy==1.26.4
scipy==1.12.0
pydantic==2.6.4
python-dotenv==1.0.1
